+++ src/App.tsx
@@ -6,6 +6,8 @@ import Formulador from '@/pages/Formulador'
 import FormuladorStep from '@/pages/FormuladorStep'
 import FormuladorConclusao from '@/pages/FormuladorConclusao'
 import Trilhas from '@/pages/Trilhas'
+import Oportunidades from '@/pages/Oportunidades'
+import Comunidade from '@/pages/Comunidade'
 
 function App() {
   return (
@@ -15,6 +17,8 @@ function App() {
           <Routes>
             <Route path="/" element={<Home />} />
             <Route path="/trilhas" element={<Trilhas />} />
+            <Route path="/oportunidades" element={<Oportunidades />} />
+            <Route path="/comunidade" element={<Comunidade />} />
             <Route path="/formulador" element={<Formulador />}>
               <Route index element={<Navigate to="identificacao" replace />} />
               <Route path="conclusao" element={<FormuladorConclusao />} />